refactor(onboard): extract inference setup seam
